feat: create functionality for filter the articles
Create functionality for filter the articles by title and category.

Filter the articles by category are handled by event click in buttonsTag element selections. When a button is clicked it will run the updateSelectedTagButton to update the button selected.

Filter the articles by title are handled by event click in search title element selection and form