feat: create a `postsSlice`
This slice for now stored a postsData, and isLoading data of the state.

Post data will get from the getPostsData async thunk function. This is where the getPostsData function will run when the App component is rendered. And accepts the argument URL which will contain the API URL to fetch the posts data.