refactor: allow the app fetch the posts data
When the app is rendered as well. The app will fetch the post data as async using the getPostsData async thunk function through the API URL stored in the .env file.

The getPostsData will accept the URL parameter and the value is based on the API URL value from the .env file.